refactor(home): update import paths for resume data
Change import paths for `ResumeData` to use the correct file name 
`resume-data` instead of `resumeData`. This ensures consistency and 
correctness in the codebase.
