+++ apps/nextjs/src/app/home/_components/Header.tsx
@@ -7,7 +7,7 @@ import TypeWriter from "typewriter-effect";
 
 import { Icon } from "@acme/ui/icons";
 
-import type { ResumeData } from "../resumeData";
+import type { ResumeData } from "../resume-data";
 import Social from "./Social";
 
 type HeaderProps = {
